move activities to an own package
