+++ app/src/main/java/es/wolfi/app/ResponseHandlers/CredentialSaveResponseHandler.java
@@ -15,7 +15,7 @@ import org.json.JSONObject;
 import java.util.Objects;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-import es.wolfi.app.passman.PasswordListActivity;
+import es.wolfi.app.passman.activities.PasswordListActivity;
 import es.wolfi.app.passman.R;
 import es.wolfi.app.passman.SettingValues;
 import es.wolfi.app.passman.SingleTon;
