+++ app/src/main/res/layout-land/content_login.xml
@@ -58,7 +58,7 @@
             android:paddingTop="@dimen/activity_vertical_margin"
             android:paddingBottom="@dimen/activity_vertical_margin"
             app:layout_behavior="@string/appbar_scrolling_view_behavior"
-            tools:context="es.wolfi.app.passman.LoginActivity"
+            tools:context="es.wolfi.app.passman.activities.LoginActivity"
             tools:showIn="@layout/activity_login">
 
             <com.google.android.material.textfield.TextInputLayout
