+++ app/src/main/res/layout/activity_login.xml
@@ -26,7 +26,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="es.wolfi.app.passman.LoginActivity">
+    tools:context="es.wolfi.app.passman.activities.LoginActivity">
 
     <include layout="@layout/content_login" />
