+++ app/src/main/res/layout/activity_password_list.xml
@@ -27,7 +27,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="es.wolfi.app.passman.PasswordListActivity">
+    tools:context="es.wolfi.app.passman.activities.PasswordListActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
