+++ app/src/main/res/layout/content_login.xml
@@ -41,7 +41,7 @@
         android:paddingRight="@dimen/activity_horizontal_margin"
         android:paddingTop="@dimen/activity_vertical_margin"
         app:layout_behavior="@string/appbar_scrolling_view_behavior"
-        tools:context="es.wolfi.app.passman.LoginActivity"
+        tools:context="es.wolfi.app.passman.activities.LoginActivity"
         tools:showIn="@layout/activity_login"
         android:orientation="vertical">
