+++ app/src/main/res/layout/content_password_list.xml
@@ -29,6 +29,6 @@
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
     tools:showIn="@layout/activity_password_list"
-    tools:context="es.wolfi.app.passman.PasswordListActivity">
+    tools:context="es.wolfi.app.passman.activities.PasswordListActivity">
 
 </FrameLayout>
