+++ app/src/main/res/menu/menu_password_list.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="es.wolfi.app.passman.PasswordListActivity" >
+    tools:context="es.wolfi.app.passman.activities.PasswordListActivity" >
     <item android:id="@+id/action_settings"
         android:title="@string/action_settings"
         android:orderInCategory="104"
